Fixed location for openjdk location
I require a fixed static location for the installation of the JDK as part of the build, i.e. if used within a maven toolchains.xml.
So create a symlink for /usr/local/openjdk-8 to the /usr/local/jdk-${JAVA_VERSION} so that toolchains.xml can just use /usr/local/openjdk-8.